refactor(core): split runtime helpers
